Extract StandardSyncPersistence from ConfigRepository
