+++ airbyte-config/config-persistence/src/main/java/io/airbyte/config/persistence/ConfigRepository.java
@@ -92,16 +92,22 @@ public class ConfigRepository {
   private final ConfigPersistence persistence;
   private final ExceptionWrappingDatabase database;
   private final ActorDefinitionMigrator actorDefinitionMigrator;
+  private final StandardSyncPersistence standardSyncPersistence;
 
   public ConfigRepository(final Database database) {
-    this(DatabaseConfigPersistence.createWithValidation(database), database, new ActorDefinitionMigrator(new ExceptionWrappingDatabase(database)));
+    this(DatabaseConfigPersistence.createWithValidation(database), database, new ActorDefinitionMigrator(new ExceptionWrappingDatabase(database)),
+        new StandardSyncPersistence(database));
   }
 
   @VisibleForTesting
-  ConfigRepository(final ConfigPersistence persistence, final Database database, final ActorDefinitionMigrator actorDefinitionMigrator) {
+  ConfigRepository(final ConfigPersistence persistence,
+                   final Database database,
+                   final ActorDefinitionMigrator actorDefinitionMigrator,
+                   final StandardSyncPersistence standardSyncPersistence) {
     this.persistence = persistence;
     this.database = new ExceptionWrappingDatabase(database);
     this.actorDefinitionMigrator = actorDefinitionMigrator;
+    this.standardSyncPersistence = standardSyncPersistence;
   }
 
   /**
@@ -408,11 +414,7 @@ public class ConfigRepository {
   }
 
   public void deleteStandardSyncDefinition(final UUID syncDefId) throws IOException {
-    try {
-      persistence.deleteConfig(ConfigSchema.STANDARD_SYNC, syncDefId.toString());
-    } catch (final ConfigNotFoundException e) {
-      LOGGER.info("Attempted to delete destination definition with id: {}, but it does not exist", syncDefId);
-    }
+    standardSyncPersistence.deleteStandardSync(syncDefId);
   }
 
   public void deleteDestinationDefinitionAndAssociations(final UUID destinationDefinitionId)
@@ -439,7 +441,7 @@ public class ConfigRepository {
         .filter(connector -> connectorDefinitionIdGetter.apply(connector).equals(definitionId))
         .collect(Collectors.toSet());
     for (final T connector : connectors) {
-      final Set<StandardSync> syncs = persistence.listConfigs(ConfigSchema.STANDARD_SYNC, StandardSync.class)
+      final Set<StandardSync> syncs = standardSyncPersistence.listStandardSync()
           .stream()
           .filter(sync -> sync.getSourceId().equals(connectorIdGetter.apply(connector))
               || sync.getDestinationId().equals(connectorIdGetter.apply(connector)))
@@ -719,15 +721,15 @@ public class ConfigRepository {
   }
 
   public StandardSync getStandardSync(final UUID connectionId) throws JsonValidationException, IOException, ConfigNotFoundException {
-    return persistence.getConfig(ConfigSchema.STANDARD_SYNC, connectionId.toString(), StandardSync.class);
+    return standardSyncPersistence.getStandardSync(connectionId);
   }
 
   public void writeStandardSync(final StandardSync standardSync) throws JsonValidationException, IOException {
-    persistence.writeConfig(ConfigSchema.STANDARD_SYNC, standardSync.getConnectionId().toString(), standardSync);
+    standardSyncPersistence.writeStandardSync(standardSync);
   }
 
   public List<StandardSync> listStandardSyncs() throws IOException, JsonValidationException {
-    return persistence.listConfigs(ConfigSchema.STANDARD_SYNC, StandardSync.class);
+    return standardSyncPersistence.listStandardSync();
   }
 
   public List<StandardSync> listStandardSyncsUsingOperation(final UUID operationId)
