+++ airbyte-config/config-persistence/src/test/java/io/airbyte/config/persistence/BaseDatabaseConfigPersistenceTest.java
@@ -49,6 +49,7 @@ class BaseDatabaseConfigPersistenceTest {
   static PostgreSQLContainer<?> container;
   static Database database;
   static DatabaseConfigPersistence configPersistence;
+  static StandardSyncPersistence standardSyncPersistence;
   static JsonSecretsProcessor jsonSecretsProcessor;
   static DataSource dataSource;
   static DSLContext dslContext;
