+++ airbyte-config/config-persistence/src/test/java/io/airbyte/config/persistence/DatabaseConfigPersistenceE2EReadWriteTest.java
@@ -7,6 +7,7 @@ package io.airbyte.config.persistence;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.spy;
 
@@ -35,6 +36,8 @@ import io.airbyte.test.utils.DatabaseConnectionHelper;
 import io.airbyte.validation.json.JsonValidationException;
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
+import org.apache.commons.lang3.NotImplementedException;
 import org.jooq.SQLDialect;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -51,6 +54,7 @@ class DatabaseConfigPersistenceE2EReadWriteTest extends BaseDatabaseConfigPersis
         ConfigsDatabaseMigrator.MIGRATION_FILE_LOCATION);
     database = new ConfigsDatabaseTestProvider(dslContext, flyway).create(false);
     configPersistence = spy(new DatabaseConfigPersistence(database));
+    standardSyncPersistence = new StandardSyncPersistence(database);
     final ConfigsDatabaseMigrator configsDatabaseMigrator =
         new ConfigsDatabaseMigrator(database, flyway);
     final DevDatabaseMigrator devDatabaseMigrator = new DevDatabaseMigrator(configsDatabaseMigrator);
@@ -79,6 +83,22 @@ class DatabaseConfigPersistenceE2EReadWriteTest extends BaseDatabaseConfigPersis
     standardActorCatalog();
     workspaceServiceAccounts();
     deletion();
+
+    checkSafeguards();
+  }
+
+  private void checkSafeguards() {
+    final String anyString = "";
+
+    // Making sure that the objects that have been migrated out of config persistence are protected with
+    // an explicit error.
+    assertThrows(NotImplementedException.class, () -> configPersistence.getConfig(ConfigSchema.STANDARD_SYNC, anyString, StandardSync.class));
+    assertThrows(NotImplementedException.class, () -> configPersistence.listConfigs(ConfigSchema.STANDARD_SYNC, StandardSync.class));
+    assertThrows(NotImplementedException.class,
+        () -> configPersistence.writeConfig(ConfigSchema.STANDARD_SYNC, anyString, MockData.standardSyncs().get(0)));
+    assertThrows(NotImplementedException.class,
+        () -> configPersistence.writeConfigs(ConfigSchema.STANDARD_SYNC, Map.of(anyString, MockData.standardSyncs().get(0))));
+    assertThrows(NotImplementedException.class, () -> configPersistence.deleteConfig(ConfigSchema.STANDARD_SYNC, anyString));
   }
 
   private void deletion() throws ConfigNotFoundException, IOException, JsonValidationException {
@@ -87,7 +107,7 @@ class DatabaseConfigPersistenceE2EReadWriteTest extends BaseDatabaseConfigPersis
       configPersistence.deleteConfig(ConfigSchema.STANDARD_WORKSPACE, standardWorkspace.getWorkspaceId().toString());
     }
     assertTrue(configPersistence.listConfigs(ConfigSchema.STANDARD_SYNC_STATE, StandardSyncState.class).isEmpty());
-    assertTrue(configPersistence.listConfigs(ConfigSchema.STANDARD_SYNC, StandardSync.class).isEmpty());
+    assertTrue(standardSyncPersistence.listStandardSync().isEmpty());
     assertTrue(configPersistence.listConfigs(ConfigSchema.STANDARD_SYNC_OPERATION, StandardSyncOperation.class).isEmpty());
     assertTrue(configPersistence.listConfigs(ConfigSchema.DESTINATION_CONNECTION, SourceConnection.class).isEmpty());
     assertTrue(configPersistence.listConfigs(ConfigSchema.STANDARD_WORKSPACE, StandardWorkspace.class).isEmpty());
@@ -146,16 +166,11 @@ class DatabaseConfigPersistenceE2EReadWriteTest extends BaseDatabaseConfigPersis
 
   private void standardSync() throws JsonValidationException, IOException, ConfigNotFoundException {
     for (final StandardSync standardSync : MockData.standardSyncs()) {
-      configPersistence.writeConfig(ConfigSchema.STANDARD_SYNC,
-          standardSync.getConnectionId().toString(),
-          standardSync);
-      final StandardSync standardSyncFromDB = configPersistence.getConfig(ConfigSchema.STANDARD_SYNC,
-          standardSync.getConnectionId().toString(),
-          StandardSync.class);
+      standardSyncPersistence.writeStandardSync(standardSync);
+      final StandardSync standardSyncFromDB = standardSyncPersistence.getStandardSync(standardSync.getConnectionId());
       assertEquals(standardSync, standardSyncFromDB);
     }
-    final List<StandardSync> standardSyncs = configPersistence
-        .listConfigs(ConfigSchema.STANDARD_SYNC, StandardSync.class);
+    final List<StandardSync> standardSyncs = standardSyncPersistence.listStandardSync();
     assertEquals(MockData.standardSyncs().size(), standardSyncs.size());
     assertThat(MockData.standardSyncs()).hasSameElementsAs(standardSyncs);
   }
